+++ backend/pyproject.toml
@@ -46,6 +46,7 @@ dependencies = [
     "fastapi[standard]>=0.115.0,<1.0.0",
     "uvicorn~=0.32",
     "redis~=5.0",
+    "tqdm==4.67.1",
 ]
 name = "bloom"
 version = "0.1.0"
